OOP changes the damn game. I'm impressed honestly!
OOP is literal game changer. I was beginning to think that functions for large modular programs were starting to become less productive and more of a hassle and then came along OOP along the way. This is wayyyy more easier for my brain! Let's swim through the rest projects now!